+++ lib/llm/providers/google.rb
@@ -18,6 +18,7 @@ module LLM
   #   ctx.talk ["Tell me about this photo", ctx.local_file("/images/photo.png")]
   #   ctx.messages.select(&:assistant?).each { print "[#{_1.role}]", _1.content, "\n" }
   class Google < Provider
+    require_relative "google/utils"
     require_relative "google/error_handler"
     require_relative "google/request_adapter"
     require_relative "google/response_adapter"
@@ -28,6 +29,7 @@ module LLM
     require_relative "google/files"
 
     include RequestAdapter
+    extend Utils
 
     HOST = "generativelanguage.googleapis.com"