+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+Changes since `v4.11.0`.
+
+## v4.11.0
+
 Changes since `v4.10.0`.
 
 ### Add